提交代码 请大神调错

+++ dangjian-admin/dangjian-admin-dao/src/main/java/cn/dlbdata/dangjian/admin/dao/mapper/PScorePartyDao.java
@@ -41,6 +41,6 @@ public interface PScorePartyDao {
 
     List<PScoreParty> getProjectScoreByUserId(@Param("userId")Integer  userId , @Param("year") Integer year);
 
-    int getSumScoreByUserId(@Param("userId")Integer  userId , @Param("year") Integer year);
+    Double getSumScoreByUserId(@Param("userId")Integer  userId , @Param("year") Integer year);
 
 }
